Updated redirections after more operations.
Updating redirections for forum operations too (creating, editing and moving - deleting and new topic, where you just go into it, should not redirect in any special way, just to the general page).
Also made it clear that the list being sent to the control panel is of forum forms (wrapper to forum class).